+++ packages/ossa/types/tabbar.d.ts
@@ -1,22 +1,22 @@
 import { ComponentClass, ReactChild } from "react";
 import OsComponent from "./base";
 
-export interface TabbarItemProps extends OsComponent {
+export interface TabBarItemProps extends OsComponent {
   icon: string | ReactChild;
   selectedIcon: string | ReactChild;
   text: string;
   useCustomIcon?: boolean;
 }
-export interface TabbarProps extends OsComponent {
-  tabsArr: Array<TabbarItemProps>;
+export interface TabBarProps extends OsComponent {
+  tabsArr: Array<TabBarItemProps>;
   isfixedBt?: boolean;
   activeColor?: string;
   defaultColor?: string;
   space?: number;
-  value?: number;
-  onClick?: (v: object) => void;
+  value: number;
+  onClick?: (v: object, index: number) => void;
 }
 
-declare const Tabbar: ComponentClass<TabbarProps>;
+declare const TabBar: ComponentClass<TabBarProps>;
 
-export { Tabbar };
+export { TabBar };