+++ packages/ossa-demo/CHANGELOG.md
@@ -1,5 +1,38 @@
 # ossa-demo
 
+## 1.4.2
+
+### Patch Changes
+
+- [#40](https://github.com/NeteaseYanxuan/OSSA/pull/40) [`4fe5d08`](https://github.com/NeteaseYanxuan/OSSA/commit/4fe5d082271600f8f48b6ce6c3b482d5165adfe1) Thanks [@CaedmonW](https://github.com/CaedmonW)! - 修复低版本 taro scss 报错以及 offset
+
+- Updated dependencies [[`4fe5d08`](https://github.com/NeteaseYanxuan/OSSA/commit/4fe5d082271600f8f48b6ce6c3b482d5165adfe1)]:
+  - ossaui@0.3.2
+
+## 1.4.1
+
+### Patch Changes
+
+- [#34](https://github.com/NeteaseYanxuan/OSSA/pull/34) [`15287c8`](https://github.com/NeteaseYanxuan/OSSA/commit/15287c84eb332af15093a67db9fa83ff2b18a8ec) Thanks [@CaedmonW](https://github.com/CaedmonW)! - when used in miniapp custom tabbar, click not work as expected
+
+- Updated dependencies [[`9fb66b9`](https://github.com/NeteaseYanxuan/OSSA/commit/9fb66b9da1898d150628941a9481a58ceff5e6bf), [`15287c8`](https://github.com/NeteaseYanxuan/OSSA/commit/15287c84eb332af15093a67db9fa83ff2b18a8ec)]:
+  - ossaui@0.3.1
+
+## 1.4.0
+
+### Minor Changes
+
+- [#24](https://github.com/NeteaseYanxuan/OSSA/pull/24) [`404bcaa`](https://github.com/NeteaseYanxuan/OSSA/commit/404bcaa15af9f1d94ca1b113632fa27b07b6e061) Thanks [@CaedmonW](https://github.com/CaedmonW)! - update taro to 3.5
+
+### Patch Changes
+
+- [#23](https://github.com/NeteaseYanxuan/OSSA/pull/23) [`ce40dce`](https://github.com/NeteaseYanxuan/OSSA/commit/ce40dce0bf6c2560ba581bb7f6560c33ac33c85c) Thanks [@hzxulin](https://github.com/hzxulin)! - fix: Update the test cases for Layout
+
+* [#32](https://github.com/NeteaseYanxuan/OSSA/pull/32) [`ec5124e`](https://github.com/NeteaseYanxuan/OSSA/commit/ec5124e6c75cc950d28745076c4dadd43f342c90) Thanks [@CaedmonW](https://github.com/CaedmonW)! - OsTabBar 类型定义更改
+
+* Updated dependencies [[`ec5124e`](https://github.com/NeteaseYanxuan/OSSA/commit/ec5124e6c75cc950d28745076c4dadd43f342c90)]:
+  - ossaui@0.3.0
+
 ## 1.3.0
 
 ### Minor Changes